Move stream library out of the kernel crate, at least until it is stabilized and generalized
