+++ capsules/src/net/mod.rs
@@ -1,1 +1,3 @@
+#[macro_use]
+pub mod stream;
 pub mod ieee802154;
