+++ capsules/src/net/stream.rs
